Change status of PEP 352 to final.
Also clarified wording to reflect the implementation; only exceptions are
new-style class.  Wording was ambiguous and suggested that you could raise
any new-style class which was considered insane by everyone at the core sprint
at PyCon 2006.  And, most importantly, a Tim-channeled Guido thought that the
current implementation was the way to go.  No one should be able to ``raise
42``.