Rewrite PEP 447 with a cleaner interface
This replaces a special method that was tuned for use by super()
by one that is also usable for PyObject_GenericGetAttr, and has
a cleaner interface (a method on the meta type instead of static method
on the type that didn't really know what it wanted to be).

Next up: provide prototype implementation, then post to python-dev again.