PEP 522: Rewrite accounting for Nathaniel's feedback
The core proposal (raising BlockingIOError) remains the same,
but the rationale and background sections now go into much more
depth for folks that want it.

The PEP now also explicitly covers the related changes to interpreter
and random module initialization, as well as providing references
to most of the related bugs.python.org messages.