A few wording fixed to pep-0484
Not sure if this is the way to suggest fixes, let me know. Reading through the pep, I found a few places where the wording/grammar is quite hard to understand at first sight; I tried to suggest a few minor changes that should improve readability while not affecting the substance of the text.