PEP 541: Make the Packaging WG the approval authority
It isn't appropriate for the PSF to ask a single community
volunteer to accept full responsibility for the PSF's
policy on PyPI name management.

Delegating the task to the Packaging Working Group is more
appropriate, since we're already responsible for handling
the PSF's budget in this area, and there's a lot of overlap
between budget management and risk management.