Fix PEP summary/email line indent in pep-0000.html (#516)
The PEP summary/email lines that aren't the first line in their group
have an extra leading space because the previous line is printed with a
trailing space. The special lines read from pep-0000.txt should be
printed with no extra ending to be consistent with what the python2
version did with a print statement that ends with a comma.