Refresh PEP-0394 recomandations for python command target
This changes the recommandation for making ``python`` command to
point to either python2 or python3 based on the official python
version of the distribution.

This solves issue where distributions avoided to create a ``python``
symlink even when they had only python3 installed, thus breaking lots
of scripts that were correctly using cross compatibility
recommandations of using the `python`.

As of 2018, is it far worse not to have a ``python`` command installed
than having it pointed to ``python3``.

One notable example are Fedora 27 or newer which has only python3
installed by default and is missing a symlink. See
https://bugzilla.redhat.com/show_bug.cgi?id=1630882