PEP 0008: Clarify self-contradictory =-spacing rules
Previously, the PEP contained roughly the following three rules, in sequence:

1. Never use spaces around `=` when used to indicate a parameter default
2. Something unrelated
3. DO use spaces around `=` when used to indicate a parameter default if there is also a parameter annotation present

This was, in my opinion, confusing, since rules 1 and 3 contradict each other. If you read both, it's not too hard to deduce the intended meaning (don't use spaces for default parameters of unannotated parameters, but do use them for annotated parameters), but it's still technically a contradiction as written, and somebody scan-reading the document could easily just see the first rule and walk away misled.

This commit attempts to clarify this part of the PEP by:
* Combining the first and third rules listed above into a single rule that addresses both annotated and unannotated parameters
* Rephrasing the first rule to indicate that it applies only to unannotated parameters, to eliminate the logical contradiction