Draft of vote details
The vote's supposed to start in a few days, so I figured it would be
good to get out ahead of any last-minute futzing with the text. Here's
a fully-fleshed-out proposal for the vote instructions and options, so
everyone can review.

I also set up a test vote on CIVS, using the text in my initial
commit, so you can see how it will actually look:
https://civs.cs.cornell.edu/cgi-bin/vote.pl?id=E_f3dd3ec110515d32&akey=add219018ca56843

I'll post these links on python-committers too.

CC: @ewdurbin