Remove misleading ".zip" example sdist (#842)
Later on in the same file, it says that the .zip example can't possibly work: "[source distributions] will be gzipped tar archives, with the .tar.gz extension. Zip archives, or other compression formats for tarballs, are not allowed at present."

This, and other documentation issues not fixed here, was noted in a blog post I was reading: https://blog.schuetze.link/2018/07/21/a-dive-into-packaging-native-python-extensions.html -- full credit to @konstin.