Add an old PEP: Concurrency Memory Model by Jeffrey Yasskin
Benjamin and I were waxing nostalgically about this and realized that
it was never submitted as a PEP draft.  Even if it's just meant to be
informational and withdrawn, I think it would be nice to have as a
historical document, and possibly as a starting point for something
more modern.

Retrieved from:
https://bazaar.launchpad.net/~jyasskin/python/memory-model-pep/view/head:/pep-memory-model.txt