create package target
the peps are built... may as well publish a tarball directly from travis.

python.org hosting is being updated and i'm trying to move it away from local
disk storage, which the current pep integration relies on.

this will allow me python.org to pull the tarball out of S3 and do it's work.