Toolchain change in the manylinux2010 i686 image
Following popular demands over at manylinux, the `i686` image was updated to use the same `devtoolset-8` as the `x86_64` image.
`devtoolset-8` was not included in the first place because was not available anywhere for `i686`. We're now rebuilding it on COPR to address concerns expressed in the manylinux repo about this software stack inconsistency.