Proposal: Ignore all empty entries within PYTHONPATH
Currently, common ways to set ``$PYTHONPATH`` also unexpectedly add the
current directory as an entry.  As a result, users have an increased
risk of unintentionally running malicious code.

This PEP proposes that all empty entries within ``$PYTHONPATH`` be ignored,
instead of being treated as the current directory.

I proposed this to <python-list@python.org>
on Thu, 27 Aug 2020 19:33:13 -0400 (EDT), with subject
"Proposal: Ignore all empty entries within $PYTHONPATH".
I received no comments of any kind. So I thought I'd propose
it here as a PEP.

Implementing this is trivial. I have created this as a PEP because
this is a behavioral change in a long-present mechanism, so there
is a relevant concern about backwards compatibility.
That said, I believe those concerns are easily addressed.
See the proposed PEP for details.

My apologies if this is the wrong process, just let me know
what you'd like me to do instead. Thank you!

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>