PEP 634: Finish up, ready for SC review (#1633)
Clean up PEP 634 (the new PEP 622).

Renamed Constant Value Pattern to Value Pattern. Renamed a few other grammar rules as well.

Leave open for the SC to decide: whether to change `name := pattern` to `pattern as name`.

Note that the companion PEP 635 (motivation and rationale) and PEP 636 (tutorial) are not yet ready.