PEP 480: Clarify automated signing solution to remove key sharing
This commit adds an additional step to the automated signing solution
to allow multiple developers to upload keys for the same package
if they are maintainers of the package. This removes the need
to share keys between developers.

An alternative would be to allow each project to delegate
to separate developers. However, this solution would require
uploading additional targets metadata which may complicate
data storage on the repository.

Signed-off-by: marinamoore <mnm678@gmail.com>