PEP 484: Mark as Final
PEP 484 is still marked as "Provisional", but as @Fidget-Spinner pointed out in python/typing#805, it shouldn't be any more. python/cpython#16204 removed talk about typing's provisional status from the module docs, and enhancements to typing have been in the form of new PEPs for the last few release cycles.

Going to mark this as a draft so Guido can have a chance to comment once he's back.