PEP 3104: Add footnote to give example (#2027)
While reading PEP 3104 to understand the ``nonlocal`` keyword, I found the explanation for why an alternative was not chosen to be vague. Further digging revealed an example for why there was a concern that ``.x`` and ``x`` could be used interchangeably. I linked this example as a footnote to save future readers time from digging to find an example. I hope this will increase comprehension of the PEP for future learners.