Do a major overhaul of the Motivation section
This is based on Pradeep's feedback that we were getting to the
point too abruptly. The quality is probably lower than the rest
of the PEP because this new writing hasn't yet had any edits
or proofreading.

Also, since I was rewriting this section anyway I went ahead
and formatted it as 80 columns instead of infinite-width. The
whole file eventually needs to be reformatted but this one section
was blocking me (it was infinite width because it started as a
google doc, but using a text editor this is not so good)