Extend the functions-as-types discussion
I wrote a google doc describing this idea in more detail so that
anyone not familiar with it can get context without it taking
up over a page of the PEP itself.

We could potentially inline that entire discussion, but I think
having a separate doc is probably better because it might trigger
a discussion of proposing that idea that isn't directly relevant
to callable type syntax.

I also added some subsubsection headers to keep this subsection
organized.