Cut back wording regarding unpacked tuple types
Leave in the discussion of how `(int, *Ts) -> bool` is equivalent
to various Callable constructions, but remove the requirement
to support an explicit `Unpack[Ts]`.

Supporting `Unpack` in new syntax shouldn't be necessary
since the new callable syntax will ship with support for the
new star syntax, so there's no need for having a solution on
earlier python versions.