PEP 585: remove mention of removal date
A definite timeline contradicts the CPython documentation, e.g. see
https://github.com/python/cpython/pull/91864

Removal will break pretty much all code written to support Python 3.8
and earlier that uses type hints. I'm not sure the one year between 3.8
going EOL and the removal date mentioned here is enough to expect all
libraries with type hints to update.

Note that the goal here isn't to decide on a removal date (or perhaps,
to decide to keep them around forever), but just to reduce fear for
Python users and to stay consistent with CPython docs.