Add Topic for release schedule PEPs (#2669)
Release PEPs are a pretty unusual kind of PEP and it's useful to me to have
them all listed together. Happy to adjust the name of the topic if desired.

I took the opportunity to adjust the status of a few release PEPs. Now all
releases for EOL versions are Final (3.5 was previously Active), and all
releases for released versions are Active (3.10 was still draft). 3.11 remains
Draft as 3.11 has not yet been released.

Co-authored-by: C.A.M. Gerlach <CAM.Gerlach@Gerlach.CAM>