PEP 495: Fix incorrect gap image (fixes #2959)
The SVG image in the "Mind the Gap" section (pep-0495-gap.svg) is incorrect. It
is essentially the same as the the SVG image in the "In the Gap" section
(pep-0495-fold.svg). This is a bug that dates back to the original commit on
2015-08-30 (814d3728115b94fbede89f5c73f5cf3065d00e7b) that uploaded the both
the PNG and SVG versions of the Fold and Gap images. The PNG version of the Gap
image is correct, but the SVG version is incorrect.

I believe this bug was invisible prior to 2022-04-23 because PEP 495 previously
referenced the PNG versions of the images (pep-0495-gap.png,
pep-0495-fold-2.png). Commit 28a5eb386a499d04c9fbaf6967ddd29842ccfaa7 deleted
the PNG images and changed the links in PEP 495 to point to the SVG versions.

The replacement SVG image attached in this PR is a manual recreation of the PNG
version of the Gap image that was deleted on 2022-04-23. It is not exactly the
same, but I think it is close enough.