PEP-11: Make unsupporting a platform less draconian
This loosens the statements about builds and configure needing to
Break and instead changes this to a mere build/configure time warning.
This is in line with how we treat platforms that are not on a support
Tier (of which there are many).  Doing anything else would basically
be unreasonable punshiment for anything in our support Tiers that then
loses its official PEP-11 support for whatever reason.