Be consistent with "key", "item", "value type" and word choices.
This follows the style of PEP 705 and the majority of the typing spec to
use "key" or "item" instead of "field" depending on the context. It also
makes use of "value type" more to be adequately more pedantic. The
special key added is now referred to as the "reserved key" to make it
clearer that it is not intended for other purposes anymore.

It also make use of type aliases like `VT` more for brevity.

Signed-off-by: Zixuan James Li <p359101898@gmail.com>