Remove [project] changes from Dependency Groups
In response to feedback, and especially to the points about ecosystem
impacts and encouraging users to move package metadata out of the
current standard location, remove support for Dependency Group
Includes in the `[project]` table.

This is now listed with rationale in Rejected Ideas, and several
sections have been updated for correctness to fit this removal.