More strongly discourage overlapping extras
In "PEP 735: Dependency Groups", the defined semantics for a potential
name conflict between an extra and a group discouraged tools from
raising errors on such conflicts.

In the revised text, this is left completely open to implementers, and
it's more clearly likely and probable that some tools will error in
these cases. Users are still discouraged from creating this situation,
regardless of tool behaviors.