Add PEP 735 note on direct extras usage
`uv` presented feedback that they allow for users to directly
reference content in the `[project]` table, even for non-package
projects. As a result, the text regarding extras as package metadata,
and the conflcit this presents with non-package projects, needs
clarification.

Simply add a section which calls this out explicitly and clarifies.