Remove mention of `requirements.in` from 735
This is extending the PEP text for an ancillary point of confusion --
there are tool users for `uv` (and possibly `pip-tools`) who do not
understand that `requirements.in` is a convention that does not
represent a significantly distinct file-format.

Discussing the difference between these only lengthens the already
very-long PEP text, and does not add value vis-a-vis community
standards. Given that neither `requirements.in` nor `requirements.txt`
have clear specifications, other than "what `pip` accepts", keeping
with the older and more standardized naming keeps things shorter and
simpler.