PEP 735: move project includes to "deferred ideas"
Move the sections on `[project]` includes and `[build-system]`
includes to Deferred Ideas, as an added section.

Restore the removed use case for this, as a subsection of one of these
Deferred Ideas.

Also, per suggestion, remove an unnecessary phrase ("defined below" in
a context where it is _immediately_ followed by the relevant
section).