PEP 396: Reclassify as Withdrawn
The ``__version__`` runtime convention hasn't been rejected,
as it's still a good way to version import packages instead of
(or in addition to) distribution packages.

We just concluded that this PEP wasn't the right vehicle for
documenting it, which is better reflected as a withdrawal rather
than as a rejection.