PEP 768: Flesh out the security design
Switch from a buffer containing Python code to a buffer containing the
path to a file containing Python code. This helps to prevent an attacker
with arbitrary memory write capabilities inside a running Python process
from also gaining arbitrary code execution if they don't already have
the ability to write to the file system.

Also, indicate that we will send a new audit event whenever this
interface is used, so that any attacker who uses it risks detection.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>