Updates to PEP-0752 including co-authoring of the PEP
The PEP-0752 is important for community driven projects - such as
Apache Airflow, not only because of security risks but also because
of concerns with public discussions about new names of distributions
following agreed and implemented pattern.

As agreed with Ofek, I am happy to become co-author of the PEP and
added more motivation explaining this case.