Remove TODO bullet point for `z` and `#` commutativity
It's better to give the formatspec one canonical ordering than permit an
overly liberal rearrangeability.

If commutativity were added, then as well as the messy description required
for the docs for the particular case of `int` data, two people could write
two different format spec that result in the same output and not realise
it or agree, because they've written different things, leading to confusion etc.